Fixed #359: Support compiling to WebAssembly using Emscripten

+++ tutorials/embree_tests/common/algorithms/parallel_partition.cpp
@@ -22,11 +22,11 @@ TEST_CASE("Test parallel_partition", "[parallel_partition")
     size_t split = std::rand() % (N+1);
 
     /* perform parallel partitioning */
-    size_t left_sum = 0, right_sum = 0;
+    uint64_t left_sum = 0, right_sum = 0;
     size_t mid = parallel_partitioning(array.data(),0,array.size(),0,left_sum,right_sum,
                                        [&] ( size_t i ) { return i < split; },
-                                       []  ( size_t& sum, unsigned v) { sum += v; },
-                                       []  ( size_t& sum, size_t v) { sum += v; },
+                                       []  ( uint64_t& sum, unsigned v) { sum += v; },
+                                       []  ( uint64_t& sum, size_t v) { sum += v; },
                                        128);
 
     /*serial_partitioning(array.data(),0,array.size(),left_sum,right_sum,
